Testimonial Unpublished Bug Fix
I forgot to add a line of code to save the testimonial after setting it
as unpublished. I used this opportunity to change the order of execution
within the original code to set the testimonial to be unpublished, if
necessary, BEFORE the initial save. Thus, removing the need to add in an
unnecessary second save. This wraps up #2.