README Update - Cron Workflow
Updated the README to reflect the changes from #4. The entire workflow
was originally designed to be kicked off by a QueueWorker, but it was
decided that, because an item would have to be placed in the worker's
queue at a interval by another cron job, that it would be redundant. So
instead, I opted to change this step from being managed by a QueueWorker
to just a regular cron job as defined in the module file. This should
wrap up #4, unless formatting needs updated.